:recycle: api: Rename pick-answer-service into pick-answer-status-service
